python/setupper/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
idna==3.10
    # via requests
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
urllib3==2.2.3
    # via requests
